Changed kioskdisable logic to skip lock task mode instead of finish().
As the kiosk browser is now more of a single purpose image, rather than
a modified version of the stock browser on an otherwise typical lollipop
installation, the stock browser was made to be the home/launcher app of
the device. This was especially necessary on 256M tailored images. This
change in behavior also required that the ro.boot.kioskdisable property
being set did not finish the activity (would infinitely reopen the app
till system crash) but rather bypass the locktaskmode.

This change primarily allows for the settings dropdown menu to become
available when androidboot.kioskdisable is set in the 'extra' env variable
of the bootloader.

Signed-off-by: Bobby Jones <rjones@gateworks.com>